+++ lib/geordi/commands/tests.rb
@@ -1,27 +1,25 @@
-desc 'tests [FILES and OPTIONS]', 'Run all employed tests'
+desc 'tests [FILES]', 'Run all employed tests'
 long_desc <<-LONGDESC
 When running `geordi tests` without any arguments, all unit tests, rspec specs
 and cucumber features will be run.
 
-When running `geordi tests` with arguments, it will invoke either
-`geordi rspec` or `geordi cucumber` to run the test, depending on whether the
-path in the first argument starts with "spec" or "features".
-
-Command line arguments are passed to `geordi rspec`/`geordi cucumber` as well.
+When passing arguments, Geordi will forward them to either `rspec` or `cucumber`,
+depending on what the first argument indicates.
 LONGDESC
 
 def tests(*args)
   if args.any?
     args, opts = Thor::Options.split(args)
+    error_message = "When passing arguments, the first argument must be either an RSpec or a Cucumber path."
 
     if args.empty?
-      Interaction.fail "geordi tests only supports arguments if one or more test files/directories are specified."
+      Interaction.fail error_message
     elsif args.first.start_with? 'spec'
       invoke 'rspec', args, opts
     elsif args.first.start_with? 'features'
       invoke 'cucumber', args, opts
     else
-      Interaction.fail "geordi tests cannot process the argument #{args.first}. The first argument has to be a path beginning with \"spec\" or \"features\"."
+      Interaction.fail error_message
     end
 
   else